scan-tool: change directory structure of request cache
Instead of using one single directory for all cached reports, the
reports are now distributed into 16 subdirectories of the cache's
root directory. This way the performance of directory listing and
similar operations does not degrade so fast for larger request
caches. It might possibly be useful to do this again later and
increase the number of directories from 16 to 256, but I will leave
that for the future. Currently, 16 directories are enough for my
use case.