move VirusTotal-specific report data to separate class
Some data members of the base class Report (as seen before this
commit) are not general information that can be seen in most of
the reports of scan APIs. They are seen in VirusTotal's reports
only, and therefore they have been moved to a separate class
named ReportVirusTotalBase. It serves as base class for ReportV2
and ReportHoneypot, both are report classes for VirusTotal APIs.
Now it should be easier to adopt the class Report as base class
for any future report classes.