scan-tool: improve rescan requests
Actually use rescan API for rescanning files instead of submitting the
whole file again. This is faster and saves bandwith. As a consequence
we also got rid of size limitations for rescans.
(No need to submit a file, so we cannot hit any size limit. The files
are already in VirusTotal's data store, because someone submitted them
earlier. Otherwise it would not be a re-scan.)