metascan: raise size limit for scans to 140 MB + use new URL
The file limit for scans with the metascan::Scanner class is raised
to 140 MB, because that is the limit on the web interface and may be
more accurate than the 50 MB we have been using until now.

The scanner (and thus the depending applications scan-tool-mso and
mso-api-request) now use the new *.metadefender.com URLs instead of
the old *.metascan-online.com URLs for scans, hash lookups and
rescans. Also all mentions of Metascan Online are replaced with
Metadefender Cloud. This follows the name change from Metascan to
Metadefender which OPSWAT put into effect on 24th February 2016.
("Old wine in new wineskins.")

Anyhow, the namespace and directories in this code are still named
"metascan", because I am too lazy to change that.