scan-tool: add option to read the API key from a file
This way the API key will not appear in the process list and/or shell
history. However, the file name can still be  seen, so proper file
permissions should be set to avoid that other users can read the API
key.

The change only affects the programs that interact with the VirusTotal
API, i.e. honeypottery, scan-tool, scan-tool-cache and vt-api-request.