scan-tool: add zlib dependency to CMake file
scan-tool depends on zlib's gzip-handling functions to build and
run properly. The depedency on zlib is implicitly given by the
dependency on libarchive (linking against libarchive will also
link against zlib, because that is a dependency of libarchive),
however that might not work as expected on some systems.
Therefore we make that dependency explicit.