mso-api-request: bump version to 1.0.0
This is most likely the last update for mso-api-request.
mso-api-request has been largely unchanged for some time now, so
I assume it to be stable, as in "it works as expected, it has all
the features that I need and I will not add new features in the
near future".