scan-tool: add possibility to ignore archive-related errors
If told via the command line parameter --ignore-extraction-errors,
the scan-tool program will ignore errors that occur during the
extraction of archive files. (The default behaviour is to exit the
program as soon as an archive-related error occurs. Use the new
option to override this behaviour. However, this might possibly
leave some temporary files and directories on disk, even after the
program has exited.)